Update HttpClient dependency and Readme.
The HttpClient library seems to use reflection,
which broke things when using this project as a
dependency for another project which used a newer HttpClient.

HypeMachine confirmed that they don't care about the fact that
they don't scrobble the current track.
Therefore I added a link to my browser extension to fix that
issue to the readme.